Fix : Change AttractionInfo Entity contentId type long to Integer
- 계획 여행지 추가에서 구현한 타입과 상이하여 수정

+++ src/main/java/com/ssafy/aroundthekorea/map/repository/AttractionInfoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.ssafy.aroundthekorea.map.domain.AttractionInfo;
 
-public interface AttractionInfoRepository extends JpaRepository<AttractionInfo, Long> {
+public interface AttractionInfoRepository extends JpaRepository<AttractionInfo, Integer> {
 	// 키워드 검색을 위한 쿼리문 생성(sido와 gugun이 null일 경우 자기자신을 반환하여 모든 sido gugun에 대해 실시)
 //	@Query("SELECT a FROM AttractionInfo a WHERE (a.title LIKE %:keyword% OR a.addr1 LIKE %:keyword% OR a.addr2 LIKE %:keyword%) AND a.sidoCode = CASE WHEN :sido IS NOT NULL THEN :sido ELSE a.sidoCode END AND a.contentTypeId = CASE WHEN :contentTypeId IS NOT NULL THEN :contentTypeId ELSE a.contentTypeId END")
 	@Query("SELECT a \r\n"
